+++ packages/define-model/src/index.ts
@@ -1,6 +1,11 @@
 import { createUnplugin } from 'unplugin'
 import { createFilter, normalizePath } from '@rollup/pluginutils'
-import { REGEX_SETUP_SFC, REGEX_VUE_SFC } from '@vue-macros/common'
+import {
+  REGEX_SETUP_SFC,
+  REGEX_VUE_SFC,
+  REGEX_VUE_SUB,
+  detectVueVersion,
+} from '@vue-macros/common'
 import { transformDefineModel } from './core'
 import {
   emitHelperCode,
@@ -9,6 +14,8 @@ import {
   useVmodelHelperCode,
   useVmodelHelperId,
 } from './core/helper'
+import type { UnpluginContextMeta } from 'unplugin'
+import type { MarkRequired } from '@vue-macros/common'
 import type { FilterPattern } from '@rollup/pluginutils'
 
 export interface Options {
@@ -27,13 +34,20 @@ export interface Options {
   unified?: boolean
 }
 
-export type OptionsResolved = Omit<Required<Options>, 'exclude'> & {
-  exclude?: FilterPattern
-}
+export type OptionsResolved = MarkRequired<
+  Options,
+  'include' | 'version' | 'unified'
+>
 
-function resolveOption(options: Options): OptionsResolved {
+function resolveOption(
+  options: Options,
+  framework: UnpluginContextMeta['framework']
+): OptionsResolved {
+  const version = options.version || detectVueVersion()
   return {
-    include: [REGEX_VUE_SFC, REGEX_SETUP_SFC],
+    include: [REGEX_VUE_SFC, REGEX_SETUP_SFC].concat(
+      version === 2 && framework === 'webpack' ? REGEX_VUE_SUB : []
+    ),
     version: 3,
     unified: true,
     ...options,
@@ -42,38 +56,45 @@ function resolveOption(options: Options): OptionsResolved {
 
 const name = 'unplugin-vue-define-model'
 
-export default createUnplugin((userOptions: Options = {}) => {
-  const options = resolveOption(userOptions)
-  const filter = createFilter(options.include, options.exclude)
+export default createUnplugin<Options | undefined, false>(
+  (userOptions = {}, { framework }) => {
+    const options = resolveOption(userOptions, framework)
+    const filter = createFilter(options.include, options.exclude)
 
-  return {
-    name,
-    enforce: 'pre',
+    return {
+      name,
+      enforce: 'pre',
 
-    resolveId(id) {
-      if (id.startsWith(helperPrefix)) return id
-    },
+      resolveId(id) {
+        if (id.startsWith(helperPrefix)) return id
+      },
 
-    loadInclude(id) {
-      return id.startsWith(helperPrefix)
-    },
+      loadInclude(id) {
+        return id.startsWith(helperPrefix)
+      },
 
-    load(_id) {
-      const id = normalizePath(_id)
-      if (id === emitHelperId) return emitHelperCode
-      else if (id === useVmodelHelperId) return useVmodelHelperCode
-    },
+      load(_id) {
+        const id = normalizePath(_id)
+        if (id === emitHelperId) return emitHelperCode
+        else if (id === useVmodelHelperId) return useVmodelHelperCode
+      },
 
-    transformInclude(id) {
-      return filter(id)
-    },
+      transformInclude(id) {
+        return filter(id)
+      },
 
-    transform(code, id) {
-      try {
-        return transformDefineModel(code, id, options.version, options.unified)
-      } catch (err: unknown) {
-        this.error(`${name} ${err}`)
-      }
-    },
+      transform(code, id) {
+        try {
+          return transformDefineModel(
+            code,
+            id,
+            options.version,
+            options.unified
+          )
+        } catch (err: unknown) {
+          this.error(`${name} ${err}`)
+        }
+      },
+    }
   }
-})
+)