+++ packages/named-template/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @vue-macros/named-template
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`d81825a`](https://github.com/sxzz/unplugin-vue-macros/commit/d81825a9bbe003f5af4ee3858241cc5bdb8f264f)]:
+  - @vue-macros/common@0.13.5
+
+## 0.1.0
+
+### Minor Changes
+
+- [#177](https://github.com/sxzz/unplugin-vue-macros/pull/177) [`934d5f8`](https://github.com/sxzz/unplugin-vue-macros/commit/934d5f8e935fc6ee0a4b3846e183b6611e10c571) Thanks [@sxzz](https://github.com/sxzz)! - Support Vite 4
+  Drop Rollup 2
+
+### Patch Changes
+
+- [`b978650`](https://github.com/sxzz/unplugin-vue-macros/commit/b978650fef0ec4d1f1c8f2bb5bf76e75b4feab26) Thanks [@sxzz](https://github.com/sxzz)! - fix reference deps
+
 ## 0.0.6
 
 ### Patch Changes