+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,58 @@
 # @vue-macros/nuxt
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - unplugin-vue-macros@1.2.2
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [[`f276a7d`](https://github.com/sxzz/unplugin-vue-macros/commit/f276a7d513cc1cbcb7dcce534e5287447b0676a3)]:
+  - unplugin-vue-macros@1.2.1
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [[`4c58923`](https://github.com/sxzz/unplugin-vue-macros/commit/4c58923f4bc78b7fce05877e428a42d3452e8d09)]:
+  - unplugin-vue-macros@1.2.0
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [[`14b5fbf`](https://github.com/sxzz/unplugin-vue-macros/commit/14b5fbf3895568665ef414f3c302c62738b7cac2)]:
+  - @vue-macros/volar@0.5.7
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - unplugin-vue-macros@1.1.1
+  - @vue-macros/short-vmodel@1.0.1
+  - @vue-macros/volar@0.5.6
+
+## 0.2.1
+
+### Patch Changes
+
+- [#177](https://github.com/sxzz/unplugin-vue-macros/pull/177) [`934d5f8`](https://github.com/sxzz/unplugin-vue-macros/commit/934d5f8e935fc6ee0a4b3846e183b6611e10c571) Thanks [@sxzz](https://github.com/sxzz)! - Support Vite 4
+  Drop Rollup 2
+- Updated dependencies [[`934d5f8`](https://github.com/sxzz/unplugin-vue-macros/commit/934d5f8e935fc6ee0a4b3846e183b6611e10c571)]:
+  - unplugin-vue-macros@1.1.0
+  - @vue-macros/volar@0.5.5
+
+## 0.2.0
+
+### Minor Changes
+
+- [`2c46f98`](https://github.com/sxzz/unplugin-vue-macros/commit/2c46f983776d741fb0dcdac5aefac5eb6829def1) Thanks [@sxzz](https://github.com/sxzz)! - support shortVmodel
+
 ## 0.1.2
 
 ### Patch Changes